move MainPage and related to enterprise/dotcom/welcome
