+++ web/src/SourcegraphWebApp.scss
@@ -240,7 +240,6 @@ hr {
 @import './global/GlobalAlerts';
 @import './global/GlobalDebug';
 @import './docSite/DocSitePage';
-@import './search/input/MainPage';
 @import './search/input/ScopePage';
 @import './search/input/SearchPage';
 @import './search/results/SearchResults';
