+++ web/src/enterprise/dotcom/welcome/CodeIntellifyBlob.tsx
@@ -5,19 +5,19 @@ import * as H from 'history'
 import * as React from 'react'
 import { Subject, Subscription } from 'rxjs'
 import { catchError, filter, map, withLatestFrom } from 'rxjs/operators'
-import { ActionItemProps } from '../../../../shared/src/actions/ActionItem'
-import { HoverMerged } from '../../../../shared/src/api/client/types/hover'
-import { ExtensionsControllerProps } from '../../../../shared/src/extensions/controller'
-import * as GQL from '../../../../shared/src/graphql/schema'
-import { getHoverActions } from '../../../../shared/src/hover/actions'
-import { HoverContext, HoverOverlay } from '../../../../shared/src/hover/HoverOverlay'
-import { getModeFromPath } from '../../../../shared/src/languages'
-import { PlatformContextProps } from '../../../../shared/src/platform/context'
-import { ErrorLike, isErrorLike } from '../../../../shared/src/util/errors'
-import { isDefined, propertyIsDefined } from '../../../../shared/src/util/types'
-import { FileSpec, ModeSpec, PositionSpec, RepoSpec, ResolvedRevSpec, RevSpec } from '../../../../shared/src/util/url'
-import { getHover } from '../../backend/features'
-import { fetchBlob } from '../../repo/blob/BlobPage'
+import { ActionItemProps } from '../../../../../shared/src/actions/ActionItem'
+import { HoverMerged } from '../../../../../shared/src/api/client/types/hover'
+import { ExtensionsControllerProps } from '../../../../../shared/src/extensions/controller'
+import * as GQL from '../../../../../shared/src/graphql/schema'
+import { getHoverActions } from '../../../../../shared/src/hover/actions'
+import { HoverContext, HoverOverlay } from '../../../../../shared/src/hover/HoverOverlay'
+import { getModeFromPath } from '../../../../../shared/src/languages'
+import { PlatformContextProps } from '../../../../../shared/src/platform/context'
+import { ErrorLike, isErrorLike } from '../../../../../shared/src/util/errors'
+import { isDefined, propertyIsDefined } from '../../../../../shared/src/util/types'
+import { FileSpec, ModeSpec, PositionSpec, RepoSpec, ResolvedRevSpec, RevSpec } from '../../../../../shared/src/util/url'
+import { getHover } from '../../../backend/features'
+import { fetchBlob } from '../../../repo/blob/BlobPage'
 
 interface Props extends ExtensionsControllerProps, PlatformContextProps {
     history: H.History
