+++ web/src/enterprise/dotcom/welcome/MainPage.tsx
@@ -3,19 +3,19 @@ import ChevronRightIcon from 'mdi-react/ChevronRightIcon'
 import CloseIcon from 'mdi-react/CloseIcon'
 import MapSearchIcon from 'mdi-react/MapSearchIcon'
 import * as React from 'react'
-import { parseSearchURLQuery } from '..'
-import { ExtensionsControllerProps } from '../../../../shared/src/extensions/controller'
-import * as GQL from '../../../../shared/src/graphql/schema'
-import { PlatformContextProps } from '../../../../shared/src/platform/context'
-import { Form } from '../../components/Form'
-import { HeroPage } from '../../components/HeroPage'
-import { PageTitle } from '../../components/PageTitle'
-import { eventLogger } from '../../tracking/eventLogger'
-import { limitString } from '../../util'
-import { queryIndexOfScope, submitSearch } from '../helpers'
+import { ExtensionsControllerProps } from '../../../../../shared/src/extensions/controller'
+import * as GQL from '../../../../../shared/src/graphql/schema'
+import { PlatformContextProps } from '../../../../../shared/src/platform/context'
+import { Form } from '../../../components/Form'
+import { HeroPage } from '../../../components/HeroPage'
+import { PageTitle } from '../../../components/PageTitle'
+import { parseSearchURLQuery } from '../../../search'
+import { queryIndexOfScope, submitSearch } from '../../../search/helpers'
+import { SearchButton } from '../../../search/input/SearchButton'
+import { eventLogger } from '../../../tracking/eventLogger'
+import { limitString } from '../../../util'
 import { CodeIntellifyBlob } from './CodeIntellifyBlob'
 import { QueryInputForModal } from './QueryInputForModal'
-import { SearchButton } from './SearchButton'
 
 interface Props extends ExtensionsControllerProps, PlatformContextProps {
     authenticatedUser: GQL.IUser | null
