+++ web/src/enterprise/dotcom/welcome/QueryInputForModal.scss
@@ -1,5 +1,3 @@
-@import './Suggestion';
-
 .query-input2-for-modal {
     width: 100%;
     position: relative;
