--- web/src/search/input/README.md
@@ -1,1 +1,0 @@
-Components for areas in the app that expose a search input
